requirements.txt
django-endless-pagination
redis
cjklib
BeautifulSoup4
django-registration
readability-lxml
django-fancy-cache
South
django_mobile
django_compressor